chore: move the `pool` golang package from to root to `internal`
Missed this previously, should be in the `internal` directory with the other packages - unless I am missing something.
